refactor: add modules to js
