+++ index.html
@@ -4,10 +4,10 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./style.css">
+    <link rel="stylesheet" href="style.css">
     <title>RSS Virtual Keyboard</title>
 </head>
 <body>
-  <script src="./js/script.js"></script>    
+  <script src="./js/script.js" type="module" ></script>    
 </body>
 </html>
