+++ js/lang.js
@@ -1,520 +1,530 @@
 const langArr = {
-    en : {
-        "Backquote" : {
-            small: '`',
-            big: '~',
-        },
-        "Digit1" : {
-            small: '1',
-            big: '!',
-        },
-        "Digit2" : {
-            small: '2',
-            big: '@',
-        },
-        "Digit3" : {
-            small: '3',
-            big: '#',
-        },
-        "Digit4" : {
-            small: '4',
-            big: '$',
-        },
-        "Digit5" : {
-            small: '5',
-            big: '%',
-        },
-        "Digit6" : {
-            small: '6',
-            big: '^',
-        },
-        "Digit7" : {
-            small: '7',
-            big: '&',
-        },
-        "Digit8" : {
-            small: '8',
-            big: '*',
-        },
-        "Digit9" : {
-            small: '9',
-            big: '(',
-        },
-        "Digit0" : {
-            small: '0',
-            big: ')',
-        },
-        "Minus" : {
-            small: '-',
-            big: '_',
-        },
-        "Equal" : {
-            small: '=',
-            big: '+',
-        },
-        "Backspace" : {
-            small: 'Backspace',
-            big: null,
-        },
-        "Tab" : {
-            small: 'Tab',
-            big: null,
-        },
-        "KeyQ" : {
-            small: 'q',
-            big: 'Q',
-        },
-        "KeyW" : {
-            small: 'w',
-            big: 'W',
-        },
-        "KeyE" : {
-            small: 'e',
-            big: 'E',
-        },
-        "KeyR" : {
-            small: 'r',
-            big: 'R',
-        },
-        "KeyT" :  {
-            small: 't',
-            big: 'T',
-        },
-        "KeyY" :  {
-            small: 'y',
-            big: 'Y',
-        },
-        "KeyU" :  {
-            small: 'u',
-            big: 'U',
-        },
-        "KeyI" :  {
-            small: 'i',
-            big: 'I',
-        },
-        "KeyO" :  {
-            small: 'o',
-            big: 'O',
-        },
-        "KeyP" :  {
-            small: 'p',
-            big: 'P',
-        },
-        "BracketLeft" :  {
-            small: '[',
-            big: '{',
-        },
-        "BracketRight" :  {
-            small: ']',
-            big: '}',
-        },
-        "Backslash" :  {
-            small: '\\',
-            big: '|',
-        },
-        "Delete" :  {
-            small: 'Del',
-            big: null,
-        },
-        "CapsLock" :  {
-            small: 'CapsLock',
-            big: 'CapsLock',
-        },
-        "KeyA" :  {
-            small: 'a',
-            big: 'A',
-        },
-        "KeyS" :  {
-            small: 's',
-            big: 'S',
-        },
-        "KeyD" :  {
-            small: 'd',
-            big: 'D',
-        },
-        "KeyF" :  {
-            small: 'f',
-            big: 'F',
-        },
-        "KeyG" : {
-            small: 'g',
-            big: 'G',
-        },
-        "KeyH" :  {
-            small: 'h',
-            big: 'H',
-        },
-        "KeyJ" :  {
-            small: 'j',
-            big: 'J',
-        },
-        "KeyK" :  {
-            small: 'k',
-            big: 'K',
-        },
-        "KeyL" :  {
-            small: 'l',
-            big: 'L',
-        },
-        "Semicolon" :  {
-            small: ';',
-            big: ':',
-        },
-        "Quote" :  {
-            small: '\'',
-            big: '"',
-        },
-        "Enter" :  {
-            small: 'Enter',
-            big: null,
-        },
-        "ShiftLeft" :  {
-            small: 'Shift',
-            big: null,
-        },
-        "KeyZ" : {
-            small: 'z',
-            big: 'Z',
-        },
-        "KeyX" : {
-            small: 'x',
-            big: 'X',
-        },
-        "KeyC" : {
-            small: 'c',
-            big: 'C',
-        },
-        "KeyV" : {
-            small: 'v',
-            big: 'V',
-        },
-        "KeyB" : {
-            small: 'b',
-            big: 'B',
-        },
-        "KeyN" : {
-            small: 'n',
-            big: 'N',
-        },
-        "KeyM" : {
-            small: 'm',
-            big: 'M',
-        },
-        "Comma" : {
-            small: ',',
-            big: '<',
-        },
-        "Period" : {
-            small: '.',
-            big: '>',
-        },
-        "Slash" : {
-            small: '/',
-            big: '?',
-        },
-        "ArrowUp" : {
-            small: '▲',
-            big: '▲',
-        },
-        "ShiftRight" : {
-            small: 'Shift',
-            big: null,
-        },
-        "ControlLeft" : {
-            small: 'Ctrl',
-            big: null,
-        },
-        "MetaLeft" : {
-            small: 'Win',
-            big: null,
-        },
-        "AltLeft" : {
-            small: 'Alt',
-            big: null,
-        },
-        "Space" : {
-            small: ' ',
-            big: ' ',
-        },
-        "AltRight" : {
-            small: 'Alt',
-            big: null,
-        },
-        "ArrowLeft" : {
-            small: '◄',
-            big: '◄',
-        },
-        "ArrowDown" : {
-            small: '▼',
-            big: '▼',
-        },
-        "ArrowRight" : {
-            small: '►',
-            big: '►',
-        },
-        "ControlRight" : {
-            small: 'Ctrl',
-            big: null,
-        },
+  en : {
+      "Backquote" : {
+          small: '`',
+          big: '~',
+      },
+      "Digit1" : {
+          small: '1',
+          big: '!',
+      },
+      "Digit2" : {
+          small: '2',
+          big: '@',
+      },
+      "Digit3" : {
+          small: '3',
+          big: '#',
+      },
+      "Digit4" : {
+          small: '4',
+          big: '$',
+      },
+      "Digit5" : {
+          small: '5',
+          big: '%',
+      },
+      "Digit6" : {
+          small: '6',
+          big: '^',
+      },
+      "Digit7" : {
+          small: '7',
+          big: '&',
+      },
+      "Digit8" : {
+          small: '8',
+          big: '*',
+      },
+      "Digit9" : {
+          small: '9',
+          big: '(',
+      },
+      "Digit0" : {
+          small: '0',
+          big: ')',
+      },
+      "Minus" : {
+          small: '-',
+          big: '_',
+      },
+      "Equal" : {
+          small: '=',
+          big: '+',
+      },
+      "Backspace" : {
+          small: 'Backspace',
+          big: null,
+      },
+      "Tab" : {
+          small: 'Tab',
+          big: null,
+      },
+      "KeyQ" : {
+          small: 'q',
+          big: 'Q',
+      },
+      "KeyW" : {
+          small: 'w',
+          big: 'W',
+      },
+      "KeyE" : {
+          small: 'e',
+          big: 'E',
+      },
+      "KeyR" : {
+          small: 'r',
+          big: 'R',
+      },
+      "KeyT" :  {
+          small: 't',
+          big: 'T',
+      },
+      "KeyY" :  {
+          small: 'y',
+          big: 'Y',
+      },
+      "KeyU" :  {
+          small: 'u',
+          big: 'U',
+      },
+      "KeyI" :  {
+          small: 'i',
+          big: 'I',
+      },
+      "KeyO" :  {
+          small: 'o',
+          big: 'O',
+      },
+      "KeyP" :  {
+          small: 'p',
+          big: 'P',
+      },
+      "BracketLeft" :  {
+          small: '[',
+          big: '{',
+      },
+      "BracketRight" :  {
+          small: ']',
+          big: '}',
+      },
+      "Backslash" :  {
+          small: '\\',
+          big: '|',
+      },
+      "Delete" :  {
+          small: 'Del',
+          big: null,
+      },
+      "CapsLock" :  {
+          small: 'CapsLock',
+          big: 'CapsLock',
+      },
+      "KeyA" :  {
+          small: 'a',
+          big: 'A',
+      },
+      "KeyS" :  {
+          small: 's',
+          big: 'S',
+      },
+      "KeyD" :  {
+          small: 'd',
+          big: 'D',
+      },
+      "KeyF" :  {
+          small: 'f',
+          big: 'F',
+      },
+      "KeyG" : {
+          small: 'g',
+          big: 'G',
+      },
+      "KeyH" :  {
+          small: 'h',
+          big: 'H',
+      },
+      "KeyJ" :  {
+          small: 'j',
+          big: 'J',
+      },
+      "KeyK" :  {
+          small: 'k',
+          big: 'K',
+      },
+      "KeyL" :  {
+          small: 'l',
+          big: 'L',
+      },
+      "Semicolon" :  {
+          small: ';',
+          big: ':',
+      },
+      "Quote" :  {
+          small: '\'',
+          big: '"',
+      },
+      "Enter" :  {
+          small: 'Enter',
+          big: null,
+      },
+      "ShiftLeft" :  {
+          small: 'Shift',
+          big: null,
+      },
+
+      "KeyZ" : {
+          small: 'z',
+          big: 'Z',
+      },
+      "KeyX" : {
+          small: 'x',
+          big: 'X',
+      },
+      "KeyC" : {
+          small: 'c',
+          big: 'C',
+      },
+      "KeyV" : {
+          small: 'v',
+          big: 'V',
+      },
+      "KeyB" : {
+          small: 'b',
+          big: 'B',
+      },
+      "KeyN" : {
+          small: 'n',
+          big: 'N',
+      },
+      "KeyM" : {
+          small: 'm',
+          big: 'M',
+      },
+      "Comma" : {
+          small: ',',
+          big: '<',
+      },
+      "Period" : {
+          small: '.',
+          big: '>',
+      },
+      "Slash" : {
+          small: '/',
+          big: '?',
+      },
+      "ArrowUp" : {
+          small: '▲',
+          big: '▲',
+      },
+      "ShiftRight" : {
+          small: 'Shift',
+          big: null,
+      },
+      "ControlLeft" : {
+          small: 'Ctrl',
+          big: null,
+      },
+      "MetaLeft" : {
+          small: 'Win',
+          big: null,
+      },
+      "AltLeft" : {
+          small: 'Alt',
+          big: null,
+      },
+      "Space" : {
+          small: ' ',
+          big: ' ',
+      },
+      "AltRight" : {
+          small: 'Alt',
+          big: null,
+      },
+      "ArrowLeft" : {
+          small: '◄',
+          big: '◄',
+      },
+      "ArrowDown" : {
+          small: '▼',
+          big: '▼',
+      },
+      "ArrowRight" : {
+          small: '►',
+          big: '►',
+      },
+      "ControlRight" : {
+          small: 'Ctrl',
+          big: null,
+      },
+      "IntlBackslash" : {
+        small: '\\',
+        big: null,
+      },
+      
+  },
+  ru : {
+    "IntlBackslash" : {
+      small: '\\',
+      big: null,
     },
-    ru : {
-        "Backquote" : {
-            small: 'ё',
-            big: 'Ё',
-        },
-        "Digit1" : {
-            small: '1',
-            big: '!',
-        },
-        "Digit2" : {
-            small: '2',
-            big: '"',
-        },
-        "Digit3" : {
-            small: '3',
-            big: '№',
-        },
-        "Digit4" : {
-            small: '4',
-            big: ';',
-        },
-        "Digit5" : {
-            small: '5',
-            big: '%',
-        },
-        "Digit6" : {
-            small: '6',
-            big: ':',
-        },
-        "Digit7" : {
-            small: '7',
-            big: '?',
-        },
-        "Digit8" : {
-            small: '8',
-            big: '*',
-        },
-        "Digit9" : {
-            small: '9',
-            big: '(',
-        },
-        "Digit0" : {
-            small: '0',
-            big: ')',
-        },
-        "Minus" : {
-            small: '-',
-            big: '_',
-        },
-        "Equal" : {
-            small: '=',
-            big: '+',
-        },
-        "Backspace" : {
-            small: 'Backspace',
-            big: null,
-        },
-        "Tab" : {
-            small: 'Tab',
-            big: null,
-        },
-        "KeyQ" : {
-            small: 'й',
-            big: 'Й',
-        },
-        "KeyW" : {
-            small: 'ц',
-            big: 'Ц',
-        },
-        "KeyE" : {
-            small: 'у',
-            big: 'У',
-        },
-        "KeyR" : {
-            small: 'к',
-            big: 'К',
-        },
-        "KeyT" :  {
-            small: 'е',
-            big: 'Е',
-        },
-        "KeyY" :  {
-            small: 'н',
-            big: 'Н',
-        },
-        "KeyU" :  {
-            small: 'г',
-            big: 'Г',
-        },
-        "KeyI" :  {
-            small: 'ш',
-            big: 'Ш',
-        },
-        "KeyO" :  {
-            small: 'щ',
-            big: 'Щ',
-        },
-        "KeyP" :  {
-            small: 'з',
-            big: 'З',
-        },
-        "BracketLeft" :  {
-            small: 'х',
-            big: 'Х',
-        },
-        "BracketRight" :  {
-            small: 'ъ',
-            big: 'Ъ',
-        },
-        "Backslash" :  {
-            small: '\\',
-            big: '/',
-        },
-        "Delete" :  {
-            small: 'Del',
-            big: null,
-        },
-        "CapsLock" :  {
-            small: 'CapsLock',
-            big: 'CapsLock',
-        },
-        "KeyA" :  {
-            small: 'ф',
-            big: 'Ф',
-        },
-        "KeyS" :  {
-            small: 'ы',
-            big: 'Ы',
-        },
-        "KeyD" :  {
-            small: 'в',
-            big: 'В',
-        },
-        "KeyF" :  {
-            small: 'а',
-            big: 'А',
-        },
-        "KeyG" : {
-            small: 'п',
-            big: 'П',
-        },
-        "KeyH" :  {
-            small: 'р',
-            big: 'Р',
-        },
-        "KeyJ" :  {
-            small: 'о',
-            big: 'О',
-        },
-        "KeyK" :  {
-            small: 'л',
-            big: 'Л',
-        },
-        "KeyL" :  {
-            small: 'д',
-            big: 'Д',
-        },
-        "Semicolon" :  {
-            small: 'ж',
-            big: 'Ж',
-        },
-        "Quote" :  {
-            small: 'э',
-            big: 'Э',
-        },
-        "Enter" :  {
-            small: 'Enter',
-            big: null,
-        },
-        "ShiftLeft" :  {
-            small: 'Shift',
-            big: null,
-        },
-        "KeyZ" : {
-            small: 'я',
-            big: 'Я',
-        },
-        "KeyX" : {
-            small: 'ч',
-            big: 'Ч',
-        },
-        "KeyC" : {
-            small: 'с',
-            big: 'С',
-        },
-        "KeyV" : {
-            small: 'м',
-            big: 'М',
-        },
-        "KeyB" : {
-            small: 'и',
-            big: 'И',
-        },
-        "KeyN" : {
-            small: 'т',
-            big: 'Т',
-        },
-        "KeyM" : {
-            small: 'ь',
-            big: 'Ь',
-        },
-        "Comma" : {
-            small: 'б',
-            big: 'Б',
-        },
-        "Period" : {
-            small: 'ю',
-            big: 'Ю',
-        },
-        "Slash" : {
-            small: '.',
-            big: ',',
-        },
-        "ArrowUp" : {
-            small: '▲',
-            big: '▲',
-        },
-        "ShiftRight" : {
-            small: 'Shift',
-            big: null,
-        },
-        "ControlLeft" : {
-            small: 'Ctrl',
-            big: null,
-        },
-        "MetaLeft" : {
-            small: 'Win',
-            big: null,
-        },
-        "AltLeft" : {
-            small: 'Alt',
-            big: null,
-        },
-        "Space" : {
-            small: ' ',
-            big: ' ',
-        },
-        "AltRight" : {
-            small: 'Alt',
-            big: null,
-        },
-        "ArrowLeft" : {
-            small: '◄',
-            big: '◄',
-        },
-        "ArrowDown" : {
-            small: '▼',
-            big: '▼',
-        },
-        "ArrowRight" : {
-            small: '►',
-            big: '►',
-        },
-        "ControlRight" : {
-            small: 'Ctrl',
-            big: null,
-        },
-    }
+      "Backquote" : {
+          small: 'ё',
+          big: 'Ё',
+      },
+      "Digit1" : {
+          small: '1',
+          big: '!',
+      },
+      "Digit2" : {
+          small: '2',
+          big: '"',
+      },
+      "Digit3" : {
+          small: '3',
+          big: '№',
+      },
+      "Digit4" : {
+          small: '4',
+          big: ';',
+      },
+      "Digit5" : {
+          small: '5',
+          big: '%',
+      },
+      "Digit6" : {
+          small: '6',
+          big: ':',
+      },
+      "Digit7" : {
+          small: '7',
+          big: '?',
+      },
+      "Digit8" : {
+          small: '8',
+          big: '*',
+      },
+      "Digit9" : {
+          small: '9',
+          big: '(',
+      },
+      "Digit0" : {
+          small: '0',
+          big: ')',
+      },
+      "Minus" : {
+          small: '-',
+          big: '_',
+      },
+      "Equal" : {
+          small: '=',
+          big: '+',
+      },
+      "Backspace" : {
+          small: 'Backspace',
+          big: null,
+      },
+      "Tab" : {
+          small: 'Tab',
+          big: null,
+      },
+      "KeyQ" : {
+          small: 'й',
+          big: 'Й',
+      },
+      "KeyW" : {
+          small: 'ц',
+          big: 'Ц',
+      },
+      "KeyE" : {
+          small: 'у',
+          big: 'У',
+      },
+      "KeyR" : {
+          small: 'к',
+          big: 'К',
+      },
+      "KeyT" :  {
+          small: 'е',
+          big: 'Е',
+      },
+      "KeyY" :  {
+          small: 'н',
+          big: 'Н',
+      },
+      "KeyU" :  {
+          small: 'г',
+          big: 'Г',
+      },
+      "KeyI" :  {
+          small: 'ш',
+          big: 'Ш',
+      },
+      "KeyO" :  {
+          small: 'щ',
+          big: 'Щ',
+      },
+      "KeyP" :  {
+          small: 'з',
+          big: 'З',
+      },
+      "BracketLeft" :  {
+          small: 'х',
+          big: 'Х',
+      },
+      "BracketRight" :  {
+          small: 'ъ',
+          big: 'Ъ',
+      },
+      "Backslash" :  {
+          small: '\\',
+          big: '/',
+      },
+      "Delete" :  {
+          small: 'Del',
+          big: null,
+      },
+      "CapsLock" :  {
+          small: 'CapsLock',
+          big: 'CapsLock',
+      },
+      "KeyA" :  {
+          small: 'ф',
+          big: 'Ф',
+      },
+      "KeyS" :  {
+          small: 'ы',
+          big: 'Ы',
+      },
+      "KeyD" :  {
+          small: 'в',
+          big: 'В',
+      },
+      "KeyF" :  {
+          small: 'а',
+          big: 'А',
+      },
+      "KeyG" : {
+          small: 'п',
+          big: 'П',
+      },
+      "KeyH" :  {
+          small: 'р',
+          big: 'Р',
+      },
+      "KeyJ" :  {
+          small: 'о',
+          big: 'О',
+      },
+      "KeyK" :  {
+          small: 'л',
+          big: 'Л',
+      },
+      "KeyL" :  {
+          small: 'д',
+          big: 'Д',
+      },
+      "Semicolon" :  {
+          small: 'ж',
+          big: 'Ж',
+      },
+      "Quote" :  {
+          small: 'э',
+          big: 'Э',
+      },
+      "Enter" :  {
+          small: 'Enter',
+          big: null,
+      },
+      "ShiftLeft" :  {
+          small: 'Shift',
+          big: null,
+      },
+      "KeyZ" : {
+          small: 'я',
+          big: 'Я',
+      },
+      "KeyX" : {
+          small: 'ч',
+          big: 'Ч',
+      },
+      "KeyC" : {
+          small: 'с',
+          big: 'С',
+      },
+      "KeyV" : {
+          small: 'м',
+          big: 'М',
+      },
+      "KeyB" : {
+          small: 'и',
+          big: 'И',
+      },
+      "KeyN" : {
+          small: 'т',
+          big: 'Т',
+      },
+      "KeyM" : {
+          small: 'ь',
+          big: 'Ь',
+      },
+      "Comma" : {
+          small: 'б',
+          big: 'Б',
+      },
+      "Period" : {
+          small: 'ю',
+          big: 'Ю',
+      },
+      "Slash" : {
+          small: '.',
+          big: ',',
+      },
+      "ArrowUp" : {
+          small: '▲',
+          big: '▲',
+      },
+      "ShiftRight" : {
+          small: 'Shift',
+          big: null,
+      },
+      "ControlLeft" : {
+          small: 'Ctrl',
+          big: null,
+      },
+      "MetaLeft" : {
+          small: 'Win',
+          big: null,
+      },
+      "AltLeft" : {
+          small: 'Alt',
+          big: null,
+      },
+      "Space" : {
+          small: ' ',
+          big: ' ',
+      },
+      "AltRight" : {
+          small: 'Alt',
+          big: null,
+      },
+      "ArrowLeft" : {
+          small: '◄',
+          big: '◄',
+      },
+      "ArrowDown" : {
+          small: '▼',
+          big: '▼',
+      },
+      "ArrowRight" : {
+          small: '►',
+          big: '►',
+      },
+      "ControlRight" : {
+          small: 'Ctrl',
+          big: null,
+      },
+  }
 }
 
 export default langArr;
