+++ style.css
@@ -17,7 +17,6 @@
 }
 .keyboard {
     width:704px;
-    height: auto;
     margin:0 auto;
     padding:5px;
     padding-left:0;
